OPE-83 relocate fixed-basis preparation
