+++ docs/reference/openghg_inversions.hbmcmc.preparation.rst
@@ -1,0 +1,7 @@
+openghg\_inversions.hbmcmc.preparation
+=======================================
+
+.. automodule:: openghg_inversions.hbmcmc.preparation
+   :members:
+   :show-inheritance:
+   :undoc-members:
