+++ newsfragments/OPE-83.misc.md
@@ -1,0 +1,1 @@
+Move the temporary fixed-basis preparation contract under ``hbmcmc`` while preserving deprecated compatibility imports from ``inversion_data``.
