+++ openghg_inversions/inversion_data/__init__.py
@@ -1,9 +1,10 @@
+import warnings
+from typing import Any
+
 from .get_data import data_processing_surface_notracer
 from .preparation import (
-    FixedBasisPreparedData,
     RhimeMergedData,
     RhimePreparedInputs,
-    prepare_fixedbasis_inversion_data,
     prepare_rhime_inputs,
 )
 from .serialise import load_merged_data, _save_merged_data
@@ -20,3 +21,19 @@ __all__ = [
     "prepare_rhime_inputs",
     "prepare_rhime_inputs_from_xarray",
 ]
+
+
+def __getattr__(name: str) -> Any:
+    """Provide warning-emitting aliases for former fixed-basis exports."""
+    if name not in {"FixedBasisPreparedData", "prepare_fixedbasis_inversion_data"}:
+        raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+    warnings.warn(
+        f"{__name__}.{name} has moved to openghg_inversions.hbmcmc.preparation; "
+        "the old import path is deprecated.",
+        FutureWarning,
+        stacklevel=2,
+    )
+    from openghg_inversions.hbmcmc import preparation as fixedbasis_preparation
+
+    return getattr(fixedbasis_preparation, name)
